+++ src/main/scala/edu/ie3/simona/config/ConfigFailFast.scala
@@ -246,6 +246,11 @@ object ConfigFailFast extends LazyLogging {
         s"Invalid time configuration." +
           s"Please ensure that the start time of the simulation is before the end time."
       )
+    if (startDate.isEqual(endDate))
+      throw new InvalidConfigParameterException(
+        s"Invalid time configuration." +
+          s"Please ensure that the start time of the simulation is not equal to the end time."
+      )
   }
 
   /** Creates a ZonedDateTime from String. If a faulty dateTime string is