+++ src/test/scala/edu/ie3/simona/model/thermal/ThermalGridSpec.scala
@@ -56,23 +56,23 @@ class ThermalGridSpec
       }
     }
 
-    "checking for required and additional demand" should {
-      "return proper information, if no required and no additional demand is apparent" in {
+    "checking for required and possible demand" should {
+      "return proper information, if no required and no possible demand is apparent" in {
         val required = MegawattHours(0d)
         val possible = MegawattHours(0d)
 
         val energyDemand = ThermalEnergyDemand(required, possible)
         energyDemand.hasRequiredDemand shouldBe false
-        energyDemand.hasAdditionalDemand shouldBe false
+        energyDemand.hasPossibleDemand shouldBe false
       }
 
-      "return proper information, if no required but additional demand is apparent" in {
+      "return proper information, if no required but possible demand is apparent" in {
         val required = MegawattHours(0d)
         val possible = MegawattHours(45d)
 
         val energyDemand = ThermalEnergyDemand(required, possible)
         energyDemand.hasRequiredDemand shouldBe false
-        energyDemand.hasAdditionalDemand shouldBe true
+        energyDemand.hasPossibleDemand shouldBe true
       }
 
       "throw exception, if required demand is higher than possible demand" in {
@@ -113,7 +113,7 @@ class ThermalGridSpec
 
         val energyDemand = ThermalEnergyDemand(required, possible)
         energyDemand.hasRequiredDemand shouldBe true
-        energyDemand.hasAdditionalDemand shouldBe true
+        energyDemand.hasPossibleDemand shouldBe true
       }
     }